Vector int, Snake selve check, game gameover, graphics drawrectvec(dim), objects, Board,

+++ Engine/HitBox.h
@@ -4,15 +4,15 @@ class HitBox
 {
 public:
 	HitBox() = default;
-	HitBox( float left_in,float right_in,float top_in,float bottem_in );
+	HitBox( int left_in,int right_in,int top_in,int bottem_in );
 	HitBox( const Vector& topleft,const Vector& bottemright );
-	HitBox( const Vector& topleft,float width,float height );
+	HitBox( const Vector& topleft,int width,int height );
 	bool isOverlappingWith( const HitBox& other ) const;
 	bool isOverlappingWith( const Vector& other ) const;
-	static HitBox fromCenter( const Vector & center,float width,float height );
+	static HitBox fromCenter( const Vector & center,int width,int height );
 public:
-	float left;
-	float right;
-	float top;
-	float bottem;
+	int left;
+	int right;
+	int top;
+	int bottem;
 };
